Moved predicate into CdkPredicate and utils methods into CdkUtils
