+++ python-checks/src/main/java/org/sonar/python/checks/cdk/DisabledRDSEncryptionCheck.java
@@ -19,13 +19,14 @@
  */
 package org.sonar.python.checks.cdk;
 
-import java.util.Locale;
 import java.util.Optional;
-import java.util.function.Predicate;
 import org.sonar.check.Rule;
 import org.sonar.plugins.python.api.SubscriptionContext;
 import org.sonar.plugins.python.api.tree.CallExpression;
-import org.sonar.plugins.python.api.tree.Expression;
+
+import static org.sonar.python.checks.cdk.CdkPredicate.isFalse;
+import static org.sonar.python.checks.cdk.CdkPredicate.isNone;
+import static org.sonar.python.checks.cdk.CdkPredicate.startsWith;
 
 @Rule(key = "S6303")
 public class DisabledRDSEncryptionCheck extends AbstractCdkResourceCheck {
@@ -48,8 +49,8 @@ public class DisabledRDSEncryptionCheck extends AbstractCdkResourceCheck {
   }
 
   protected void checkDatabaseArguments(SubscriptionContext ctx, CallExpression resourceConstructor) {
-    Optional<ExpressionTrace> argEncrypted = getArgument(ctx, resourceConstructor, ARG_ENCRYPTED);
-    Optional<ExpressionTrace> argEncryptionKey = getArgument(ctx, resourceConstructor, ARG_ENCRYPTION_KEY);
+    Optional<CdkUtils.ExpressionTrace> argEncrypted = CdkUtils.getArgument(ctx, resourceConstructor, ARG_ENCRYPTED);
+    Optional<CdkUtils.ExpressionTrace> argEncryptionKey = CdkUtils.getArgument(ctx, resourceConstructor, ARG_ENCRYPTION_KEY);
 
     if (argEncrypted.isEmpty() && argEncryptionKey.isEmpty()) {
       ctx.addIssue(resourceConstructor.callee(), DB_OMITTING_MESSAGE);
@@ -66,19 +67,15 @@ public class DisabledRDSEncryptionCheck extends AbstractCdkResourceCheck {
   }
 
   protected void checkCfnDatabaseArguments(SubscriptionContext ctx, CallExpression resourceConstructor) {
-    getArgument(ctx, resourceConstructor, ARG_ENCRYPTED).ifPresentOrElse(
+    CdkUtils.getArgument(ctx, resourceConstructor, ARG_ENCRYPTED).ifPresentOrElse(
       argumentTrace -> argumentTrace.addIssueIf(isFalse(), UNENCRYPTED_MESSAGE),
       () -> ctx.addIssue(resourceConstructor.callee(), CFNDB_OMITTING_MESSAGE)
     );
   }
 
   protected boolean isEngineAurora(SubscriptionContext ctx, CallExpression resourceConstructor) {
-    return getArgument(ctx, resourceConstructor, "engine")
+    return CdkUtils.getArgument(ctx, resourceConstructor, "engine")
       .filter(argumentTrace -> argumentTrace.hasExpression(startsWith("aurora"))).isPresent();
-  }
-
-  protected Predicate<Expression> startsWith(String expected) {
-    return e -> getStringValue(e).filter(str -> str.toLowerCase(Locale.ROOT).startsWith(expected)).isPresent();
   }
 }
